misc: Add some random, half completed stuff

+++ include/kpl/random.h
@@ -1,3 +1,5 @@
 #pragma once
 
-bool kpl_rand(void* buffer, ssize_t buffer_size);
+#include "kpl/common.h"
+
+bool kpl_rand(void* buffer, i32 buffer_size);
