+++ src/kpl_random.c
@@ -1,7 +1,8 @@
 #include "kpl/random.h"
+
 #include <sys/random.h>
 
-bool kpl_rand(void* buffer, ssize_t buffer_size)
+bool kpl_rand(void* buffer, i32 buffer_size)
 {
     const int flags = 0;
     if (getrandom(buffer, buffer_size, flags) == -1)
